- completed attribute binary value accessors (some optimizations are still required) - added a lot of comments to experimental attribute binary representation structs
git-svn-id: https://svn-eos.cs.ovgu.de/repos/staff/mschulze/Research/MIKRO/famouso@1577 235a714a-1e17-0410-b3bb-947d69a96898

+++ tst/experiments/Attributes/BinaryRepresentation/Utils.h
@@ -135,7 +135,7 @@ void print(const uint8_t* array, uint8_t length) {
 
 template <typename Attr>
 void printCase() {
-	::logging::log::emit() << "Selected case " << CaseSelector<Attr, uint16_t, 1, 2, 3, 4, 5, 6>::value << ::logging::log::endl;
+	::logging::log::emit() << "Selected case " << static_cast<uint16_t>(CaseSelector<Attr, uint16_t, 1, 2, 3, 4, 5, 6>::value) << ::logging::log::endl;
 }
 
 #endif
